feat: rename id only macro
