+++ mysk-lib-macros/src/fetch_level_variant/mod.rs
@@ -1,1 +1,2 @@
 pub mod impl_fetch_level_variant_from;
+pub mod impl_id_only_variant_from;
