+++ mysk-lib/src/models/teacher/fetch_levels/id_only.rs
@@ -1,4 +1,4 @@
-use mysk_lib_macros::{id_only_variant_boiler_plate, impl_fetch_level_variant_from};
+use mysk_lib_macros::{impl_fetch_level_variant_from, impl_id_only_variant_from};
 use serde::{Deserialize, Serialize};
 use sqlx::PgPool;
 use uuid::Uuid;
@@ -11,4 +11,4 @@ pub struct IdOnlyTeacher {
     pub id: Uuid,
 }
 
-id_only_variant_boiler_plate!(IdOnlyTeacher, DbTeacher);
+impl_id_only_variant_from!(IdOnlyTeacher, DbTeacher);
